% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodygram.R
\name{bodygram}
\alias{bodygram}
\title{Bodygrams}
\usage{
bodygram(data = data, model = "futuristic_male")
}
\arguments{
\item{data}{A dataframe. The dataframe with all the coordinates points. Must be equal to \code{data}.}

\item{model}{A character string. One of the available models. Check model_types(data = data) if not sure. It defaults to 'futuristic_male

@return A data frame:
\describe{
\item{Id}{Numeric. Unique id of each muscle area.}
\item{View}{A character String. The type of view (Anterior, Posterior, Left or Right)}
\item{Part}{A character string. Upper or Lower Body.}
\item{Group}{A character String. The name of the muscle group.}
\item{Muscle}{A character String. The name of the muscle}
\item{Side}{A character String. Whether it it left or right side of the body}
\item{x}{A number. x coordinates}
\item{y}{A number. y coordinates}
}}
}
\description{
This function returns a dataframe with the coordinates of a unique body diagram.
}
\examples{
bodygram(data = data)

}
