% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bodygram.R
\name{geom_bdgramr}
\alias{geom_bdgramr}
\title{geom_bdgramr}
\usage{
geom_bdgramr(
  mapping = NULL,
  data = NULL,
  position = "identity",
  stat = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetics, the x,y and grouping part of the aesthetics}

\item{data}{A dataframe. The dataframe with all the coordinates points. Usually the return of \code{bdgramr()}}

\item{position}{The position parameter. It defaults to "identity"}

\item{stat}{The stats parameter. Itdefaults to "identity"}

\item{na.rm}{Whether to remove NAs in the dataset. It deffaults to FALSE.}

\item{show.legend}{Whether to show leged in the plot. It defaults to NA}

\item{inherit.aes}{Whether to inherit the aesthetics. It defaults to TRUE}

\item{...}{Any possible extra arguments}
}
\value{
A ggproto object
}
\description{
This function creates a custom geom for ggplot.
}
\examples{
model_df <- bdgramR::bdgramr(data = data, model = "futuristic_male")
plot <- ggplot2::ggplot(data = model_df, ggplot2::aes(x, y, group = Id)) +
         geom_bdgramr(lty = 3, color = "white", ggplot2::aes(fill = Group))
plot

}
