% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bodygram.R
\name{geom_bodygram}
\alias{geom_bodygram}
\title{Bodygram polygons}
\usage{
geom_bodygram(
  mapping = NULL,
  data = NULL,
  position = "identity",
  stat = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetics, the x,y and grouping part of the aesthetics}

\item{data}{A dataframe. Input data from the athlete. If \code{NULL}, then a reference bodygram is plotted.}

\item{position}{The position parameter. It defaults to "identity"}

\item{stat}{The stats parameter. Itdefaults to "identity"}

\item{na.rm}{Whether to remove NAs in the dataset. It deffaults to FALSE.}

\item{show.legend}{Whether to show leged in the plot. It defaults to NA}

\item{inherit.aes}{Whether to inherit the aesthetics. It defaults to TRUE}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[=layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[=draw_key]{key glyphs}, to change the display of the layer in the legend.
}}
}
\description{
Visualize bodygrams for reference or with athlete data.
}
\examples{
bodygram_data <- bodygram()
bodygram_plot <- ggplot(data = bodygram_data, aes(x, y, group = Id)) +
  geom_bodygram(lty = 3, color = "black", aes(fill = Group))
bodygram_plot

}
