% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{models}
\alias{models}
\title{Full dataset with coordinate data for all models}
\format{
Data frame with 49,603 rows 9 columns
\describe{
\item{Model}{A character string. The type of body diagram.}
\item{Id}{Numeric. Unique id of each muscle area.}
\item{View}{A character String. The type of view (Anterior, Posterior, Left or Right)}
\item{Part}{A character string. Upper or Lower Body.}
\item{Group}{A character String. The name of the muscle group.}
\item{Muscle}{A character String. The name of the muscle}
\item{Side}{A character String. Whether it it left or right side of the body}
\item{x}{A number. x coordinates}
\item{y}{A number. y coordinates}

}
}
\usage{
models
}
\description{
A data frame containing all the information from all model types
}
\keyword{datasets}
