% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodygram.R
\name{bodygram}
\alias{bodygram}
\title{Returns a dataframe with the coordinates of a unique body diagram}
\usage{
bodygram(data = models, model = "futuristic_male")
}
\arguments{
\item{data}{A dataframe. Input data from athlete. Default produces a generic body diagram.}

\item{model}{A character string. One of the available models. Call model_types(data = data) for full list. Default is 'futuristic_male}
}
\value{
A data frame:
\describe{
\item{Id}{Numeric. Unique id of each muscle area.}
\item{View}{A character String. The type of view (Anterior, Posterior, Left or Right)}
\item{Part}{A character string. Upper or Lower Body.}
\item{Group}{A character String. The name of the muscle group.}
\item{Muscle}{A character String. The name of the muscle}
\item{Side}{A character String. Whether it it left or right side of the body}
\item{x}{A number. x coordinates}
\item{y}{A number. y coordinates}
}
}
\description{
Outputs a dataframe of selected body diagram. Use to render the desired body diagram.
}
\examples{
bodygram()

}
