% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_types.R
\name{model_types}
\alias{model_types}
\title{Body Diagram Model Types}
\usage{
model_types()
}
\arguments{
\item{data}{A dataframe. The dataframe with all the coordinates points.}
}
\value{
A data frame:
\describe{
\item{Model}{A factor. A column displaying all the available diagram model types}
}
}
\description{
provides a list with the names of all the available diagram's models.
}
\examples{
model_types()

}
